# [GSoC] OpenCV.js: WASM SIMD optimization 2.0