--- a/modules/js/perf/perf_imgproc/perf_blur.js
+++ b/modules/js/perf/perf_imgproc/perf_blur.js
@@ -25,28 +25,29 @@ function perf() {
     }
     let totalCaseNum, currentCaseId;
 
-    const GaussianBlurSize = [cvSize.szODD, cvSize.szQVGA, cvSize.szVGA, cvSize.sz720p];
-    const GaussianBlurType = ["CV_8UC1", "CV_8UC4", "CV_16UC1", "CV_16SC1", "CV_32FC1"];
+    const BlurSize = [cvSize.szODD, cvSize.szQVGA, cvSize.szVGA, cvSize.sz720p];
+    const Blur5x16Size = [cvSize.szVGA, cvSize.sz720p];
+    const BlurType = ["CV_8UC1", "CV_8UC4", "CV_16UC1", "CV_16SC1", "CV_32FC1"];
+    const BlurType5x5 = ["CV_8UC1", "CV_8UC4", "CV_16UC1", "CV_16SC1", "CV_32FC1", "CV_32FC3"];
     const BorderType3x3 = ["BORDER_REPLICATE", "BORDER_CONSTANT"];
-    const BorderType3x3ROI = ["BORDER_REPLICATE", "BORDER_CONSTANT", "BORDER_REFLECT", "BORDER_REFLECT101"];
+    const BorderTypeAll = ["BORDER_REPLICATE", "BORDER_CONSTANT", "BORDER_REFLECT", "BORDER_REFLECT101"];
 
-    const combiGaussianBlurBorder3x3 = combine(GaussianBlurSize, GaussianBlurType, BorderType3x3);
-    const combiGaussianBlurBorder3x3ROI = combine(GaussianBlurSize, GaussianBlurType, BorderType3x3ROI);
+    const combiBlur3x3 = combine(BlurSize, BlurType, BorderType3x3);
+    const combiBlur16x16 = combine(Blur5x16Size, BlurType, BorderTypeAll);
+    const combiBlur5x5 = combine(Blur5x16Size, BlurType5x5, BorderTypeAll);
 
-    function addGaussianBlurCase(suite, type) {
-        suite.add('gaussianBlur', function() {
-            cv.GaussianBlur(src, dst, ksize, 1, 0, borderType);
+    function addBlurCase(suite, type) {
+        suite.add('blur', function() {
+            cv.blur(src, dst, ksize, new cv.Point(-1,-1), borderType);
           }, {
               'setup': function() {
                 let size = this.params.size;
                 let matType = cv[this.params.matType];
                 let borderType = cv[this.params.borderType];
-                let type = this.params.type;
-                
-                let src = new cv.Mat(size, matType);
-                let dst = new cv.Mat(size, matType);
                 let ksizeNum = this.params.ksize;
                 let ksize = new cv.Size(ksizeNum, ksizeNum);
+                let src = new cv.Mat(size, matType);
+                let dst = new cv.Mat(size, matType);
                 },
               'teardown': function() {
                 src.delete();
@@ -55,16 +56,16 @@ function perf() {
           });
     }
 
-    function addGaussianBlurModeCase(suite, combination, type) {
+    function addBlurModeCase(suite, combination, type) {
       totalCaseNum += combination.length;
       for (let i = 0; i < combination.length; ++i) {
         let size =  combination[i][0];
         let matType = combination[i][1];
         let borderType = combination[i][2];
-        let ksizeArray = [3, 5];
-        
+        let ksizeArray = [3, 16, 5];
+
         let params = {size: size, matType:matType, ksize: ksizeArray[type], borderType:borderType};
-        addKernelCase(suite, params, type, addGaussianBlurCase);
+        addKernelCase(suite, params, type, addBlurCase);
       }
     }
 
@@ -79,24 +80,25 @@ function perf() {
           paramObjs.push({name:"size", value:"", reg:[""], index:0});
           paramObjs.push({name:"matType", value:"", reg:["/CV\_[0-9]+[FSUfsu]C[0-9]/"], index:1});
           paramObjs.push({name:"borderMode", value: "", reg:["/BORDER\_\\w+/"], index:2});
-          let locationList = decodeParams2Case(params, paramObjs,gaussianBlurCombinations);
+          let locationList = decodeParams2Case(params, paramObjs,blurCombinations);
 
           for (let i = 0; i < locationList.length; i++){
               let first = locationList[i][0];
               let second = locationList[i][1];
-              addGaussianBlurModeCase(suite, [gaussianBlurCombinations[first][second]], first);
+              addBlurModeCase(suite, [blurCombinations[first][second]], first);
             }
       } else {
         log("no filter or getting invalid params, run all the cases");
-        addGaussianBlurModeCase(suite, combiGaussianBlurBorder3x3, 0);
-        addGaussianBlurModeCase(suite, combiGaussianBlurBorder3x3ROI, 1);
+        addBlurModeCase(suite, combiBlur3x3, 0);
+        addBlurModeCase(suite, combiBlur16x16, 1);
+        addBlurModeCase(suite, combiBlur5x5, 2);
       }
-      setBenchmarkSuite(suite, "gaussianBlur", currentCaseId);
-      log(`Running ${totalCaseNum} tests from gaussianBlur`);
+      setBenchmarkSuite(suite, "blur", currentCaseId);
+      log(`Running ${totalCaseNum} tests from blur`);
       suite.run({ 'async': true }); // run the benchmark
   }
 
-    let gaussianBlurCombinations = [combiGaussianBlurBorder3x3, combiGaussianBlurBorder3x3ROI];
+    let blurCombinations = [combiBlur3x3, combiBlur16x16, combiBlur5x5];
 
     if (isNodeJs) {
       const args = process.argv.slice(2);