… modules/core/include/opencv2/core/hal/intrin_wasm.hpp
… modules/js/CMakeLists.txt
… modules/js/perf/perf_64bits.html
… modules/js/perf/perf_imgproc/perf_blur.html
… modules/js/perf/perf_imgproc/perf_cvtcolor.js
… modules/js/perf/perf_imgproc/perf_dilate.html
… modules/js/perf/perf_imgproc/perf_erode.html
… modules/js/perf/perf_imgproc/perf_erode.js
… modules/js/perf/perf_imgproc/perf_filter2D.html
… modules/js/perf/perf_imgproc/perf_filter2D.js
… modules/js/perf/perf_imgproc/perf_gaussianBlur.html
… modules/js/perf/perf_imgproc/perf_gaussianBlur.js
… modules/js/perf/perf_imgproc/perf_medianBlur.html
… modules/js/perf/perf_imgproc/perf_medianBlur.js
… modules/js/perf/perf_imgproc/perf_pyrDown.html
… modules/js/perf/perf_imgproc/perf_pyrDown.js
… modules/js/perf/perf_imgproc/perf_remap.html
… modules/js/perf/perf_imgproc/perf_remap.js
… modules/js/perf/perf_imgproc/perf_resize.js
… modules/js/perf/perf_imgproc/perf_scharr.html
… modules/js/perf/perf_imgproc/perf_scharr.js
… modules/js/perf/perf_imgproc/perf_sobel.html
… modules/js/perf/perf_imgproc/perf_sobel.js
… modules/js/perf/perf_imgproc/perf_threshold.js
… modules/js/perf/perf_imgproc/perf_warpAffine.html
… modules/js/perf/perf_imgproc/perf_warpAffine.js
… modules/js/perf/perf_imgproc/perf_warpPerspective.html
… modules/js/perf/perf_imgproc/perf_warpPerspective.js
… modules/js/src/loader.js
… platforms/js/build_js.py