## Overview

* Proposal: [OpenCV.js: WASM SIMD optimization 2.0](https://docs.google.com/document/d/1a59OL-gmlyywRYw2Dui07u_DvsuqqlL9GuBr52hA9YU/edit)
* Mentor: Vitaly Tuzov @terfendail, Ningxin Hu @huningxin,
* Student: Kun Liang @lionkunonly

<cut/>

### This pull request changes

1. Update the version of supported emscripten to 1.39.16 and modify the js files in test and perf test to fit it. Add required emscripten version into opencv.js build instructions.<br>
2. Implement the type 64 intrinsic and add perf test for some kernels.<br>
3. Optimize the perf test. Remove the redundant code in the perf test and move the functions that play similar roles into the file `perf_helpfunc.js`. Reconstruct current perf tests.<br>
4. Add more perf tests for kernels. Supported perf tests: `cvtColor`, `resize`, `threshold`, `Sobel`, `filter2D`, `Scharr`, `gaussianBlur`, `blur`, `medianBlur`, `erode`, `dilate`, `remap`, `warpAffine`, `warpPerspective`, `pyrDown`.<br>
5. Implement a loader that can detect the features of the browser and load the corresponding version of OpenCV.js automatically. It utilizes the help of [WebAssembly Feature Detection](https://github.com/GoogleChromeLabs/wasm-feature-detect).<br>

### The Test
Test Environment:
```
OS: Ubuntu Linux 18.04.4
Emscripten: 1.39.16, LLVM upstream backend
Browser: Chrome, Version 85.0.4183.26 dev (64-bit)
Hardware: Intel(R) Core(TM) i7-8700 CPU @ 3.20GHz with 12 logical cores
```
* OpenCV.js tests: all passed
* Universal Intrinsics WASM backend test: all passed
* All perf tests: all run successfully

### Results
* Perf test for 64bit intrinsics
```
OS: Ubuntu Linux 18.04.4
Emscripten: 1.39.16, LLVM upstream backend
Browser: Chrome, Version 85.0.4183.26 dev (64-bit)
Hardware: Intel(R) Core(TM) i7-8700 CPU @ 3.20GHz with 12 logical cores
```

|  Function  | Before 64bit implementation (ms) |  After 64bit implementation (ms) | Speedup |
|---|---|---|---|
| countNonZero()  | 0.5386 | 0.5221  | 1.032x |
| Mat::dot() | 0.8079 | 0.7960 | 1.015x |
| split()  | 2.1213 | 2.0483 |  1.036x |
| merge() | 2.2383 | 2.2264 | 1.005x  |

The performance of kernels after 64-bit implementation is similar to the performance before implementation.

* Performance of resize kernel with widen instructions
```
OS: Ubuntu Linux 18.04.4
Emscripten: 1.39.16, LLVM upstream backend
Browser: Chrome, Version 85.0.4183.26 dev (64-bit)
Hardware: Intel(R) Core(TM) i7-8700 CPU @ 3.20GHz with 12 logical cores
```
|  Parameters   | Mean time of Scalar | Mean time of SIMD with `shift`(ms)| Speedup | Mean time of SIMD with `widen`(ms) | Speedup |
|---|---|---|---| ---|---|
| (CV_8UC4, 1280x720, 640x480)  | 5.113 | 2.888  | 1.77x | 4.031  | 1.27x |
| (CV_8UC1, 1280x720, 640x480) | 1.870 | 1.128 | 1.66x  | 1.249 | 1.50x |

The perform of the implementation with `widen` instructions does not bring improvement. One of the reasons is that the `wasm_v8x16_shuffle` and `widen` instructions have to be used together. The original implementation is left finally.

* Perf tests

|  Supported Kernels now (lines)  | Supported Kernels before (lines) |
|---|---|
| cvtColor (421)  | cvtColor (572) |
| resize (165) | resize (262) |
| threshold (158)  | threshold (217) |
| Sobel (170) | - |
| filter2D (127) | - |
| Scharr (156) | - |
| gaussianBlur (126) | - |
| blur (130) | - |
| medianBlur (118) | - |
| erode (117) | - |
| dilate (117) | - |
| remap (182) | - |
| warpAffine (130) | - |
| warpPerspective (143) | - |
| pyrDown (116) | - |

**The collected performance data is stored in the Google Drive.** [Performance data](https://docs.google.com/spreadsheets/d/1lVPdlLLSET1WT7yJZquSLcBgDcVtaExAbFk6Boxwjoo/edit?usp=sharing)

* [Final Report](https://gist.github.com/lionkunonly/bbb50a839faf7acd1543f142d45254f6)
* [Perf tests and loader demo](https://lionkunonly.github.io/opencvjs/index.html)(May need the latest version of Chrome-Dev)
* [Video](https://youtu.be/fUQofzvh7mM)

### Performance Analysis

Because there are more than ten kernels are tested and some of them are tested with parameters with different data type and channels, the table that records the performance data is large. So I put the performance data in my personal Google drive.

[Performance data](https://docs.google.com/spreadsheets/d/1lVPdlLLSET1WT7yJZquSLcBgDcVtaExAbFk6Boxwjoo/edit?usp=sharing)

Analysis 
Based on the collected performance data, the SIMD optimization works as we expected in most situations. It can achieve similar performance as the SSE2 optimization for the native OpenCV. For example, for the kernel `blur` with the parameter `(1280x720, CV_8UC1, BORDER_REPLICATE) ksize=3`, the SIMD version has `1.357x`  speed up and the SSE2 has `1.415x` speed up. Sometimes the SIMD optimization is better. For example, for the kernel `pyrDown` with the parameter `(1920x1080, CV_32FC4)`, the SIMD version has `3.094x` speed up and the SSE2 has `1.83` speed up.  The data that achieves similar or better speed up is tagged with the color green.

However, there still exist some bad cases, for example, for the kernel `blur` with `(1280x720, CV_32FC1, BORDER_REPLICATE) ksize=3`, the SIMD version has `0.519` speed up which is 2x slower than the SSE optimization for native OpenCV. Such data is tagged with the color yellow.

The data tagged with color red means that it is unnormal. In my point of view, some yellow data is brought by the red unnormal data like the kernel `medianBlur` with the parameter `(1280x720, CV_16SC1, 5)`.

I hope the collected data can help people know the gap between the SIMD optimization in the OpenCV.js and the optimization in the native OpenCV better.

```
force_builders_only=linux,docs,Custom
buildworker:Custom=linux-4
build_image:Docs=docs-js
build_image:Custom=javascript-simd:1.39.16
```