@lionkunonly 2020-08-25
Hi, I have implemented the OpenCV.js loader and modified the `js_setup.markdown` to explain the usage of the loader. But I am not sure my expression is clear enough. So, Cloud you give some suggestions?