@lionkunonly 2020-08-27
@huningxin @terfendail Hi, I have modified the implementation based on your comments in code review. However, the most important change in this implementation is that I use the [WebAssembly Feature Detection](https://github.com/GoogleChromeLabs/wasm-feature-detect) in my implementation. The reason for using it is that I find the original implementation may fail when the browser update in the future. With the help of the [WebAssembly Feature Detection](https://github.com/GoogleChromeLabs/wasm-feature-detect) would be available even if the browser update. Because this library would update at that time.