@lionkunonly 2020-08-28
> @lionkunonly , could you please share the tests and the results you did for your loader?

### The Scene1: SIMD and threads are enabled.
#### The path configuration of the example
![Config1](https://user-images.githubusercontent.com/26929797/91514375-2b695900-e919-11ea-8378-cca7c0d5fc73.PNG)
#### The log for loading status and html elements
![result1](https://user-images.githubusercontent.com/26929797/91514410-4b008180-e919-11ea-9eca-918ce5d76a9b.PNG)
![result2](https://user-images.githubusercontent.com/26929797/91514412-4cca4500-e919-11ea-8fdd-f3fdf42f4992.PNG)

**In this scene, the application load the opencv.js with simd and threads optimization successfully.**


### The Scene2: SIMD and threads are enabled, but the path to threads+simd version is invalid.
#### The path configuration in js file
![Config2](https://user-images.githubusercontent.com/26929797/91514709-f8739500-e919-11ea-974a-43ce6bfca1d5.PNG)
##### Results
![result3](https://user-images.githubusercontent.com/26929797/91514752-23f67f80-e91a-11ea-918e-393c3dcd468f.PNG)
![result4](https://user-images.githubusercontent.com/26929797/91514754-2527ac80-e91a-11ea-972f-24d310041fdd.PNG)

**In this scene, the path for opencv.js with threads and simd optimization is empty. So the loader loads the opencv.js with simd optimization automatically.**

### The Scene3: Threads is enabled, simd is diabled.
#### The feature configuration in browser
![Config3](https://user-images.githubusercontent.com/26929797/91515010-c3b40d80-e91a-11ea-9c25-b9efd95cdff3.PNG)
##### Results
![result5](https://user-images.githubusercontent.com/26929797/91515020-cadb1b80-e91a-11ea-9898-913fa074b93d.PNG)
![result6](https://user-images.githubusercontent.com/26929797/91515022-cd3d7580-e91a-11ea-9155-1004ec82938b.PNG)

**In this scene, the loader loads the opencv.js with threads optimization automatically without any failed log like the log in the scene2.**


Finally, I am trying to design a demo to show the effect of the loader and deploy the demo on the github.io.