@lionkunonly 2020-08-30
@terfendail @terfendail   The final report is [Here](https://gist.github.com/lionkunonly/bbb50a839faf7acd1543f142d45254f6). You can leave your comment in the [Google drive version](https://docs.google.com/document/d/1cZ9vbZca7uhwTpzo1mNYfbBkfWAPF-iFo6rbNQph4uo/edit?usp=sharing) .