@huningxin 2020-08-31
> You can try the example with the Demo -- [Perf tests and loader demo](https://lionkunonly.github.io/opencvjs/index.html)

The [loader demo](https://lionkunonly.github.io/opencvjs/loaderShow/index.html) works great regarding to my test, even when [disabling WASM in Chrome](https://github.com/stevespringett/disable-webassembly#disabling-guidance). Thanks for that! 

My another suggestion would be using `gaussianBlur` kernel that has good threads speedup (4.98X) and simd speedup (3.36X). So user can see difference easily.

> The final report is [Here](https://gist.github.com/lionkunonly/bbb50a839faf7acd1543f142d45254f6). You can leave your comment in the [Google drive version](https://docs.google.com/document/d/1cZ9vbZca7uhwTpzo1mNYfbBkfWAPF-iFo6rbNQph4uo/edit?usp=sharing) .

Please add the summary and analysis of your performance data into the final report. Both of them are valuable. Thanks!