@huningxin 2020-08-31
> Should I remove the `[WIP]` in this PR title?

It seems all tasks are done. Please remove that so the OpenCV maintainers can review.