@lionkunonly 2020-09-04
@terfendail @huningxin  Hi, Vitaly and Ningxin. The video report has been upload to YouTube. You can check it [Here](https://youtu.be/FT6jQLfZdHE).